move ocr types to own file
